+++ internal/storage/gcs/bucket.go
@@ -32,8 +32,14 @@ const (
 	NonHierarchical
 	Hierarchical
 	Unknown
+)
+
+const (
 	// ReqIdField is the key for the value of
 	// GCS req-id in context.
+	// This is used by debugBucket
+	// for passing down Request ID
+	// into the underlying bucket implementation.
 	ReqIdField string = "GcsReqId"
 )
