--- tools/integration_tests/log_content/small_file_upload_log_test.go
@@ -1,36 +1,0 @@
-// Copyright 2023 Google Inc. All Rights Reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package log_content
-
-import (
-	"testing"
-
-	"github.com/googlecloudplatform/gcsfuse/v2/tools/integration_tests/util/setup"
-)
-
-const (
-	SmallFileSize            = MiB
-	DirForSmallFileUploadLog = "dirForSmallFileUploadLog"
-)
-
-func TestSmallFileUploadFileLog(t *testing.T) {
-	uploadFile(t, DirForBigFileUploadLog, SmallFileSize)
-
-	// The file being uploaded is too small (<16 MB) for progress logs
-	// to be printed.
-	unexpectedLogSubstrings := []string{"bytes uploaded so far"}
-	logString := extractRelevantLogsFromLogFile(t, setup.LogFile(), &logFileOffset)
-	verifyUnexpectedSubstringsInLogs(t, logString, unexpectedLogSubstrings)
-}
